Release swc-plugin v0.7.0 (swc_core v16.2.2 / Next.js 15.2.0)
This commit also sets `lto = true`, as the underlying issue appears to
have been fixed.

Note that according to <https://plugins.swc.rs/>, Next.js 15.2.0 is only
compatible with `swc_core` <15.0.0, however, I've tested the plugin with
Next.js 15.2.0 without issues, and figured we might as well be on the
latest version if it seems to work.

Also update the README file to clarify a bit about what this fork is and
does. Here's basically what I do:

- Open up `Cargo.toml` and change the version for `swc_core`.
- Run `npm run build-wasm`.
- Copy the resulting `swc_plugin.wasm` into a Next.js project.
- Build the Next.js project. If it works, it works.

If everything is good, I bump the version and release a new version.